make product insert a bulk upsert
When calling the Celery task to download offers for a product,
it can save a lot of offers. And from performace point of view,
it does not make much sense to save the offers one by one. It also does
not make much sense to try to figure out if the product is already saved
(ie we should do an update) or if it does not exist in our database at
all (ie we should do an insert). Some databases offer an "upsert" - try
to insert and if there's a conflict (primary key already exists), then
try to do an update. So that's what we are using here.